Issue #92 Run go test on fs package as non-root
